rename: pages folder -> screens folder
