+++ mobile/__tests__/screens/HomeScreen.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react-native';
-import Home from '../../src/pages/HomeScreen';
+import Home from '../../src/screens/HomeScreen';
 
 describe('Home', () => {
   it('renders without crashing', () => {
