+++ mobile/__tests__/screens/SettingScreen.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react-native';
-import Setting from '../../src/pages/SettingScreen';
+import Setting from '../../src/screens/SettingScreen';
 
 describe('Setting', () => {
   it('renders without crashing', () => {
